[Refactor] Refactor CCE nodepools update
